project/GUI/requirements.txt
annoy==1.17.2
beautifulsoup4==4.12.2
fastai==2.7.12
Flask==2.2.3
flask_compress==1.13
ipython==8.12.0
ipywidgets==8.0.6
matplotlib==3.7.1
nbconvert==7.3.1
nbformat==5.8.0
numpy==1.24.2
packaging==23.1
parameterized==0.9.0
Pillow==9.5.0
Pillow==9.5.0
protobuf==3.20.3
scikit_learn==1.2.2
setuptools==67.6.1
streamlit==1.21.0
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.5.0
